pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "alife-thesis-sim"
version = "0.1.0"
description = "Research-grade artificial life simulation framework for emergent evolutionary dynamics."
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "matplotlib>=3.8",
    "PyYAML>=6.0",
    "pygame>=2.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"

